Profile bugfixes done
