--- client/src/actions/profileActions.js
@@ -1,41 +1,0 @@
-import axios from "axios";
-import { GET_PROFILE, PROFILE_LOADING, GET_ERRORS } from "./types";
-
-export const getCurrentProfile = () => dispatch => {
-  dispatch(setProfileLoading());
-  axios
-    .get("/api/user/current")
-    .then(res =>
-      dispatch({
-        type: GET_PROFILE,
-        payload: res.data
-      })
-    )
-    .catch(err =>
-      dispatch({
-        type: GET_PROFILE,
-        payload: {}
-      })
-    );
-};
-
-export const updateProfile = (data, history) => dispatch => {
-  axios
-    .post(`/api/user/update/${data.id}`, data)
-    .then(res => {
-      history.push("/");
-    })
-    .catch(err =>
-      dispatch({
-        type: GET_ERRORS,
-        payload: err.response.data
-      })
-    );
-};
-
-// Profile loading
-export const setProfileLoading = () => {
-  return {
-    type: PROFILE_LOADING
-  };
-};
