+++ client/src/reducers/authReducer.js
@@ -1,6 +1,9 @@
 import isEmpty from "../validation/is-empty";
-
-import { SET_CURRENT_USER } from "../actions/types";
+import {
+  SET_CURRENT_USER,
+  GET_PROFILE,
+  PROFILE_LOADING
+} from "../actions/types";
 
 const initialState = {
   isAuthenticated: false,
@@ -14,6 +17,17 @@ export default function(state = initialState, action) {
         ...state,
         isAuthenticated: !isEmpty(action.payload),
         user: action.payload
+      };
+    case PROFILE_LOADING:
+      return {
+        ...state,
+        loading: true
+      };
+    case GET_PROFILE:
+      return {
+        ...state,
+        user: action.payload,
+        loading: false
       };
     default:
       return state;
