--- client/src/reducers/profileReducer.js
@@ -1,22 +1,0 @@
-import { GET_PROFILE, PROFILE_LOADING } from "../actions/types";
-const initialState = {
-  profile: null,
-  loading: false
-};
-export default function(state = initialState, action) {
-  switch (action.type) {
-    case PROFILE_LOADING:
-      return {
-        ...state,
-        loading: true
-      };
-    case GET_PROFILE:
-      return {
-        ...state,
-        profile: action.payload,
-        loading: false
-      };
-    default:
-      return state;
-  }
-}
